multi: Remove unnecessary convs found by unconvert.
This removes all unnecessary typecast conversions as found by the
unconvert linter.

+++ blockchain/chain.go
@@ -843,7 +843,8 @@ func LockTimeToSequence(isSeconds bool, locktime uint32) uint32 {
 	// shift the locktime over by 9 since the time granularity is in
 	// 512-second intervals (2^9). This results in a max lock-time of
 	// 33,553,920 seconds, or 1.1 years.
-	return wire.SequenceLockTimeIsSeconds | uint32(locktime>>wire.SequenceLockTimeGranularity)
+	return wire.SequenceLockTimeIsSeconds |
+		locktime>>wire.SequenceLockTimeGranularity
 }
 
 // getReorganizeNodes finds the fork point between the main chain and the passed
