+++ database/ffldb/blockio.go
@@ -760,7 +760,7 @@ func newBlockStore(basePath string, network wire.BitcoinNet) *blockStore {
 		writeCursor: &writeCursor{
 			curFile:    &lockableFile{},
 			curFileNum: uint32(fileNum),
-			curOffset:  uint32(fileOff),
+			curOffset:  fileOff,
 		},
 	}
 	store.openFileFunc = store.openFile
